跟新Mode以便后续更新 添加新的Http请求 async XXXX 添加Base62

+++ Cloudreve-API-DLL/Json/User/上传文件Json.cs
@@ -87,7 +87,7 @@
             /// <summary>
             /// 文件大小
             /// </summary>
-            public int chunkSize { get; set; }
+            public long chunkSize { get; set; }
             /// <summary>
             /// 时间戳
             /// </summary>
